requirements.txt
numpy==1.25.2
Pillow==10.0.0
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
typing_extensions==4.7.1
